implement array parameter setting in the gui. fix state removal when done in init - sewer and catchment add state in init -> need to remove in deinit - fix pumpingstation init - fix parameters of catchment - adapt flowfuns to new catchment parameters

+++ src/cd3core/flow.cpp
@@ -149,7 +149,7 @@ void Flow::dump() const {
 	cout << endl;
 }
 
-unsigned int Flow::countUnits(Flow::CalculationUnit unit) {
+size_t Flow::countUnits(Flow::CalculationUnit unit) {
 	cd3assert(fd.defined, "flow not defined");
 	return fd.unit_names[unit].size();
 }
